Move link to resume
I moved it to the bottom of the Work section. I did this because I feel like it makes more sense to include the resume next to projects I've worked on so they can learn more about my work history if they're interested. Work and education history isn't listed on my website, but it is on the resume.