Remove "Clear Form" button on contact form
I don't see a lot of value in being able to clear the form with a button click. I don't think there would be many cases where a user would want to clear the form. If they did, there are only 4 fields; it would be easy to clear manually.

Having the "Clear Form" button just seems like a risk to me. I could see a user accidentally clicking the button and losing the message they took the time to write up.