feature: Remove Cloud Directory from portfolio
Cloud Directory is very similar to Cloud Accounting in a lot of ways, and the work I did was very similar. I don't know that having both is adding a lot of value.

Plus, I never finished Cloud Directory, and I didn't do a ton of work on it. I could keep it, but I think the portfolio would be better without it.