feature: Improve Constant Improvement value section
The line about not wanting to be a halfway decent developer always
seemed like it didn't have quite the right tone; it seemed to suggestion
an unhealthy aspiration to be something special; that's not actually
how I feel or what I want to say. Removing that line doesn't remove
any real information (I communicate the same thing in other lines),
and it drops the questionable tone from the section.

I think the section feels much better without it.