feature: Don't describe myself as a strictly back-end developer
While I definitely prefer to be on the back-end and that's where I'm
strongest, I also don't want to give people the impression that I am
unwilling to do front-end work or that I can't contribute in that area.
It just depends on what kind of front-end work is needed (Hotwire
FTW!) and how much of it is needed.