feature: Improve Work section
I trimmed down the projects quite a bit. Previously, I was trying to give
a comprehensive list of all the major projects I've worked on. I think
that's excessive though. I think it makes more sense to just pull out a
select few that gives a nice overview of the type of work I've done.

I decided to remove the images for each project because I didn't have
a good image for some of them (so I was just using a company logo),
and I'm not sure the "good" images were providing much value anyways.